[Sdk] move model generation to SDK #22
