+++ src/generators/sdk/model/entity-views-generation.ts
@@ -1,7 +1,7 @@
-import {Entity, EntityAttribute, ProjectModel, View} from "./cuba-model";
+import {Entity, EntityAttribute, ProjectModel, View} from "../../../common/model/cuba-model";
 import * as ts from "typescript";
 import {LiteralTypeNode} from "typescript";
-import {collectAttributesFromHierarchy} from "../utils";
+import {collectAttributesFromHierarchy} from "./model-utils";
 
 const VIEW_NAME_TYPE_SUFFIX = 'ViewName';
 const VIEW_NAME_TYPE_PARAMETER = 'V';
