+++ src/generators/sdk/services/queries-generation.ts
@@ -1,13 +1,13 @@
-import {RestQuery} from "../model/cuba-model";
+import {RestQuery} from "../../../common/model/cuba-model";
 import {collectMethods, createMethodParamsType, MethodWithOverloads} from "./method-params-type";
 import * as ts from "typescript";
 import {PropertyAssignment, TypeAliasDeclaration} from "typescript";
-import {createIncludes, importDeclaration, ImportInfo} from "../import-utils";
+import {createIncludes, importDeclaration, ImportInfo} from "../../../common/import-utils";
 import {CreateItemResult, cubaAppCallFunc} from "./services-generation";
-import {CUBA_APP_MODULE_SPEC, CUBA_APP_TYPE, FETCH_OPTIONS_TYPE} from "../constants";
-import {renderTSNodes} from "../ts-helpers";
-import {exportModifier} from "../ts-shorthands";
-import {ModelContext} from "../utils";
+import {CUBA_APP_MODULE_SPEC, CUBA_APP_TYPE, FETCH_OPTIONS_TYPE} from "../../../common/constants";
+import {renderTSNodes} from "../../../common/ts-helpers";
+import {exportModifier} from "../../../common/ts-shorthands";
+import {ModelContext} from "../model/model-utils";
 
 const QUERIES_VAR_NAME = 'restQueries';
