+++ src/generators/sdk/services/services-generation.ts
@@ -1,4 +1,4 @@
-import {RestService} from "../model/cuba-model";
+import {RestService} from "../../../common/model/cuba-model";
 import * as ts from "typescript";
 import {
   ArrowFunction,
@@ -9,12 +9,12 @@ import {
   TypeAliasDeclaration,
   VariableStatement
 } from "typescript";
-import {renderTSNodes} from "../ts-helpers";
+import {renderTSNodes} from "../../../common/ts-helpers";
 import {collectMethods, createMethodParamsType, MethodWithOverloads} from "./method-params-type";
-import {createIncludes, importDeclaration, ImportInfo} from "../import-utils";
-import {CUBA_APP_MODULE_SPEC, CUBA_APP_NAME, CUBA_APP_TYPE, FETCH_OPTIONS_NAME, FETCH_OPTIONS_TYPE} from "../constants";
-import {exportModifier, idn, param, str} from "../ts-shorthands";
-import {ModelContext} from "../utils";
+import {createIncludes, importDeclaration, ImportInfo} from "../../../common/import-utils";
+import {CUBA_APP_MODULE_SPEC, CUBA_APP_NAME, CUBA_APP_TYPE, FETCH_OPTIONS_NAME, FETCH_OPTIONS_TYPE} from "../../../common/constants";
+import {exportModifier, idn, param, str} from "../../../common/ts-shorthands";
+import {ModelContext} from "../model/model-utils";
 
 const REST_SERVICES_VAR_NAME = 'restServices';
